# app-swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/engine/users": {
      "get": {
        "operationId": "EngineUsersController_findMany",
        "parameters": [
          { "name": "curPage", "required": false, "in": "query", "schema": { "type": "number" } },
          { "name": "perPage", "required": false, "in": "query", "schema": { "type": "number" } },
          { "name": "searchText", "required": false, "in": "query", "schema": { "type": "string" } },
          { "name": "sort", "required": false, "in": "query", "schema": { "type": "string" } },
          { "name": "projectId", "required": false, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": { "schema": { "$ref": "#/components/schemas/FindManyEngineUserResponse" } }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/EngineError" },
                    { "$ref": "#/components/schemas/ValidationError" }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Engine"]
      }
    },
    "/api/engine/users/{id}": {
      "put": {
        "operationId": "EngineUsersController_updateOne",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UpdateEngineUserDto" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineUserDto" } } }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/EngineError" },
                    { "$ref": "#/components/schemas/ValidationError" }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Engine"]
      },
      "get": {
        "operationId": "EngineUsersController_findOne",
        "parameters": [{ "name": "id", "required": true, "in": "path", "schema": { "type": "string" } }],
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/EngineUserDto" } } }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/EngineError" },
                    { "$ref": "#/components/schemas/ValidationError" }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Engine"]
      }
    },
    "/api/engine/users/send-invitation-links": {
      "post": {
        "operationId": "EngineUsersController_sendInvitationLinks",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/SendInvitationLinksArgs" } } }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/StatusResponse" } } }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    { "$ref": "#/components/schemas/EngineError" },
                    { "$ref": "#/components/schemas/ValidationError" }
                  ]
                }
              }
            }
          }
        },
        "tags": ["Engine"]
      }
    },
...
    "/api/engine/oauth/verification": {
      "post": {
        "operationId": "EngineOAuthController_oauthVerification",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": { "schema": { "$ref": "#/components/schemas/EngineOAuthVerificationArgs" } }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/TokensResponse" } } }
          }
        },
        "tags": ["Engine"]
      }
    },
    "/api/engine/oauth/google": {
      "get": {
        "operationId": "EngineGoogleOAuthController_googleAuth",
        "parameters": [
          { "name": "redirect_uri", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "client_id", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Engine"]
      }
    },
    "/api/engine/oauth/google/redirect": {
      "get": {
        "operationId": "EngineGoogleOAuthController_googleAuthRedirect",
        "parameters": [
          { "name": "redirect_uri", "required": true, "in": "query", "schema": { "type": "string" } },
          { "name": "client_id", "required": true, "in": "query", "schema": { "type": "string" } }
        ],
        "responses": { "200": { "description": "" } },
        "tags": ["Engine"]
      }
    },
...
  },
  "info": { "title": "", "description": "", "version": "1.0.0", "contact": {} },
...
  "servers": [],
  "components": {
    "securitySchemes": { "bearer": { "scheme": "bearer", "bearerFormat": "JWT", "type": "http" } },
    "schemas": {
...
      "StatusResponse": {
        "type": "object",
        "properties": { "message": { "type": "string" } },
        "required": ["message"]
      },
...
      "FindManyResponseMeta": {
        "type": "object",
        "properties": {
          "curPage": { "type": "number" },
          "perPage": { "type": "number" },
          "totalResults": { "type": "number" }
        },
        "required": ["totalResults"]
      },
...
      "EngineRefreshSession": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "userAgent": { "type": "string", "nullable": true },
          "userIp": { "type": "string", "nullable": true },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "userData": { "type": "object", "nullable": true },
          "enabled": { "type": "boolean" },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "EngineProject": { "$ref": "#/components/schemas/EngineProject" },
          "EngineUser": { "$ref": "#/components/schemas/EngineUser" }
        },
        "required": [
          "id",
          "userAgent",
          "userIp",
          "expiresAt",
          "userData",
          "enabled",
          "userId",
          "projectId",
          "createdAt",
          "updatedAt"
        ]
      },
      "EngineProject": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "nameLocale": { "type": "object", "nullable": true },
          "clientId": { "type": "string" },
          "clientSecret": { "type": "string" },
          "public": { "type": "boolean" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "EngineEmailTemplate": { "type": "array", "items": { "$ref": "#/components/schemas/EngineEmailTemplate" } },
          "EngineOAuthToken_EngineOAuthToken_projectIdToEngineOAuthProvider": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EngineOAuthToken" }
          },
          "EngineRefreshSession": { "type": "array", "items": { "$ref": "#/components/schemas/EngineRefreshSession" } },
          "EngineUser": { "type": "array", "items": { "$ref": "#/components/schemas/EngineUser" } }
        },
        "required": ["id", "name", "nameLocale", "clientId", "clientSecret", "public", "createdAt", "updatedAt"]
      },
...
      "EngineOAuthProvider": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "name": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "EngineOAuthProviderSettings": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/EngineOAuthProviderSettings" }
          },
          "EngineOAuthToken": { "type": "array", "items": { "$ref": "#/components/schemas/EngineOAuthToken" } }
        },
        "required": ["id", "name", "createdAt", "updatedAt"]
      },
      "EngineOAuthToken": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "grantedAt": { "type": "string", "format": "date-time" },
          "expiresAt": { "type": "string", "format": "date-time", "nullable": true },
          "tokenType": { "type": "string", "nullable": true },
          "scope": { "type": "string", "nullable": true },
          "verificationCode": { "type": "string", "nullable": true },
          "userId": { "type": "string" },
          "projectId": { "type": "string" },
          "providerId": { "type": "string" },
          "providerUserId": { "type": "string" },
          "providerUserData": { "type": "object", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "EngineOAuthProvider_EngineOAuthToken_projectIdToEngineOAuthProvider": {
            "$ref": "#/components/schemas/EngineProject"
          },
          "EngineOAuthProvider": { "$ref": "#/components/schemas/EngineOAuthProvider" },
          "EngineUser": { "$ref": "#/components/schemas/EngineUser" }
        },
        "required": [
          "id",
          "grantedAt",
          "expiresAt",
          "tokenType",
          "scope",
          "verificationCode",
          "userId",
          "projectId",
          "providerId",
          "providerUserId",
          "providerUserData",
          "createdAt",
          "updatedAt"
        ]
      },
      "EngineUser": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "username": { "type": "string", "nullable": true },
          "roles": { "type": "string", "nullable": true },
          "firstname": { "type": "string", "nullable": true },
          "lastname": { "type": "string", "nullable": true },
          "gender": { "type": "string", "nullable": true },
          "birthdate": { "type": "string", "format": "date-time", "nullable": true },
          "picture": { "type": "string", "nullable": true },
          "appData": { "type": "object", "nullable": true },
          "revokedAt": { "type": "string", "format": "date-time", "nullable": true },
          "emailVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "phoneVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "timezone": { "type": "number", "format": "float", "nullable": true },
          "lang": { "type": "string", "nullable": true },
          "projectId": { "type": "string" },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "EngineOAuthToken": { "type": "array", "items": { "$ref": "#/components/schemas/EngineOAuthToken" } },
          "EngineRefreshSession": { "type": "array", "items": { "$ref": "#/components/schemas/EngineRefreshSession" } },
          "EngineProject": { "$ref": "#/components/schemas/EngineProject" }
        },
        "required": [
          "id",
          "email",
          "phone",
          "username",
          "roles",
          "firstname",
          "lastname",
          "gender",
          "birthdate",
          "picture",
          "appData",
          "revokedAt",
          "emailVerifiedAt",
          "phoneVerifiedAt",
          "timezone",
          "lang",
          "projectId",
          "createdAt",
          "updatedAt"
        ]
      },
      "TokensResponse": {
        "type": "object",
        "properties": {
          "accessToken": { "type": "string" },
          "refreshToken": { "type": "string" },
          "user": { "$ref": "#/components/schemas/EngineUser" }
        },
        "required": ["accessToken", "refreshToken", "user"]
      },
...
      "EngineUserDto": {
        "type": "object",
        "properties": {
          "id": { "type": "string" },
          "email": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "username": { "type": "string", "nullable": true },
          "roles": { "type": "string", "nullable": true },
          "firstname": { "type": "string", "nullable": true },
          "lastname": { "type": "string", "nullable": true },
          "gender": { "type": "string", "nullable": true },
          "birthdate": { "type": "string", "format": "date-time", "nullable": true },
          "picture": { "type": "string", "nullable": true },
          "appData": { "type": "object", "nullable": true },
          "revokedAt": { "type": "string", "format": "date-time", "nullable": true },
          "emailVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "phoneVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "timezone": { "type": "number", "format": "float", "nullable": true },
          "lang": { "type": "string", "nullable": true },
          "createdAt": { "type": "string", "format": "date-time" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "email",
          "phone",
          "username",
          "roles",
          "firstname",
          "lastname",
          "gender",
          "birthdate",
          "picture",
          "appData",
          "revokedAt",
          "emailVerifiedAt",
          "phoneVerifiedAt",
          "timezone",
          "lang",
          "createdAt",
          "updatedAt"
        ]
      },
...
      "FindManyEngineUserResponse": {
        "type": "object",
        "properties": {
          "engineUsers": { "type": "array", "items": { "$ref": "#/components/schemas/EngineUserDto" } },
          "meta": { "$ref": "#/components/schemas/FindManyResponseMeta" }
        },
        "required": ["engineUsers", "meta"]
      },
      "UpdateEngineUserDto": {
        "type": "object",
        "properties": {
          "email": { "type": "string" },
          "phone": { "type": "string", "nullable": true },
          "username": { "type": "string", "nullable": true },
          "password": { "type": "string" },
          "roles": { "type": "string", "nullable": true },
          "firstname": { "type": "string", "nullable": true },
          "lastname": { "type": "string", "nullable": true },
          "gender": { "type": "string", "nullable": true },
          "birthdate": { "type": "string", "format": "date-time", "nullable": true },
          "picture": { "type": "string", "nullable": true },
          "appData": { "type": "object", "nullable": true },
          "revokedAt": { "type": "string", "format": "date-time", "nullable": true },
          "emailVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "phoneVerifiedAt": { "type": "string", "format": "date-time", "nullable": true },
          "timezone": { "type": "number", "format": "float", "nullable": true },
          "lang": { "type": "string", "nullable": true }
        }
      },
      "SendInvitationLinksArgs": {
        "type": "object",
        "properties": { "emails": { "type": "string" } },
        "required": ["emails"]
      },
...
      "EngineOAuthVerificationArgs": {
        "type": "object",
        "properties": { "verificationCode": { "type": "string" }, "fingerprint": { "type": "string" } },
        "required": ["verificationCode", "fingerprint"]
      },
...
      "EngineErrorEnum": {
        "type": "string",
        "enum": [
          "ENGINE-000",
          "ENGINE-001",
          "ENGINE-002",
          "ENGINE-003",
          "ENGINE-004",
          "ENGINE-005",
          "ENGINE-006",
          "ENGINE-007",
          "ENGINE-008",
          "ENGINE-009",
          "ENGINE-010",
          "ENGINE-011",
          "ENGINE-012",
          "ENGINE-013",
          "ENGINE-014",
          "ENGINE-015",
          "ENGINE-016",
          "ENGINE-017",
          "ENGINE-018"
        ]
      },
      "EngineError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Engine error (ENGINE-000), User not found (ENGINE-001), Wrong password (ENGINE-002), User is exists (ENGINE-003), Wrong activate email code (ENGINE-004), Activate email not processed (ENGINE-005), Activate email processed (ENGINE-006), Refresh token not provided (ENGINE-007), Session expired (ENGINE-008), Invalid refresh session (ENGINE-009), Access token expired (ENGINE-010), User is exists (ENGINE-011), Email not verified (ENGINE-012), Forbidden (ENGINE-013), Wrong old password (ENGINE-014), Non-existent role specified (ENGINE-015), Bad access token (ENGINE-016), Your session has been blocked (ENGINE-017), Verification code not found (ENGINE-018)",
            "example": "Engine error"
          },
          "code": { "example": "ENGINE-000", "allOf": [{ "$ref": "#/components/schemas/EngineErrorEnum" }] },
          "metadata": { "type": "object" }
        },
        "required": ["message", "code"]
      },
...
      "ValidationErrorEnum": { "type": "string", "enum": ["VALIDATION-000"] },
      "ValidationErrorMetadataConstraint": {
        "type": "object",
        "properties": { "name": { "type": "string" }, "description": { "type": "string" } },
        "required": ["name", "description"]
      },
      "ValidationErrorMetadata": {
        "type": "object",
        "properties": {
          "property": { "type": "string" },
          "constraints": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/ValidationErrorMetadataConstraint" }
          },
          "children": { "type": "array", "items": { "$ref": "#/components/schemas/ValidationErrorMetadata" } }
        },
        "required": ["property", "constraints"]
      },
      "ValidationError": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string",
            "description": "Validation error (VALIDATION-000)",
            "example": "Validation error"
          },
          "code": { "example": "VALIDATION-000", "allOf": [{ "$ref": "#/components/schemas/ValidationErrorEnum" }] },
          "metadata": { "type": "array", "items": { "$ref": "#/components/schemas/ValidationErrorMetadata" } }
        },
        "required": ["message", "code"]
      },
...
    }
  }
}